Fix following merge by extracting JsonUtils to WithJsonHelper
