+++ src/main/java/se/vandmo/dependencylock/maven/json/DependenciesLockFileJson.java
@@ -16,15 +16,13 @@ import se.vandmo.dependencylock.maven.Dependency;
 import se.vandmo.dependencylock.maven.LockFileAccessor;
 import se.vandmo.dependencylock.maven.ProfiledDependencies;
 
-public final class DependenciesLockFileJson {
+public final class DependenciesLockFileJson extends WithJsonHelper {
   private static final String V3 = "3";
 
   private final LockFileAccessor dependenciesLockFile;
-  private final VersionConstraintJsonSerializer versionConstraintJsonSerializer;
 
   private DependenciesLockFileJson(LockFileAccessor dependenciesLockFile) {
     this.dependenciesLockFile = dependenciesLockFile;
-    this.versionConstraintJsonSerializer = new VersionConstraintJsonSerializer();
   }
 
   public static DependenciesLockFileJson from(LockFileAccessor dependenciesLockFile) {
@@ -40,45 +38,36 @@ public final class DependenciesLockFileJson {
     }
   }
 
-   public void write(ProfiledDependencies dependencies) {
-        JsonNodeFactory jsonNodeFactory = JsonNodeFactory.instance;
-        ObjectNode json = jsonNodeFactory.objectNode();
-        json.put("version", V3);
-        json.set("artifacts", JsonUtils.buildArtifactsJson(dependencies.artifacts(), jsonNodeFactory));
-        json.set(
-                "dependencies",
-                JsonUtils.buildDependenciesJson(dependencies.getSharedDependencies(), jsonNodeFactory));
-        json.set(
-                "profiles", JsonUtils.buildProfilesJson(dependencies.profileEntries(), jsonNodeFactory));
-        try (Writer writer = dependenciesLockFile.writer()) {
-            writeJson(writer, json);
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
-        }
+  public void write(ProfiledDependencies dependencies) {
+    JsonNodeFactory jsonNodeFactory = JsonNodeFactory.instance;
+    ObjectNode json = jsonNodeFactory.objectNode();
+    json.put("version", V3);
+    json.set("artifacts", buildArtifactsJson(dependencies.artifacts(), jsonNodeFactory));
+    json.set(
+        "dependencies",
+        buildDependenciesJson(dependencies.getSharedDependencies(), jsonNodeFactory));
+    json.set("profiles", buildProfilesJson(dependencies.profileEntries(), jsonNodeFactory));
+    try (Writer writer = dependenciesLockFile.writer()) {
+      writeJson(writer, json);
+    } catch (IOException e) {
+      throw new UncheckedIOException(e);
+    }
   }
 
-    private JsonNode toJson(LockedDependencies lockedDependencies, JsonNodeFactory nodeFactory) {
+  private JsonNode toJson(Dependencies lockedDependencies, JsonNodeFactory nodeFactory) {
     ObjectNode json = nodeFactory.objectNode();
     json.set("dependencies", asJson(lockedDependencies, nodeFactory));
     return json;
   }
 
-  private JsonNode asJson(LockedDependencies lockedDependencies, JsonNodeFactory nodeFactory) {
-    ArrayNode json = nodeFactory.arrayNode();
-    for (Dependency lockedDependency : lockedDependencies.lockedEntities) {
-      json.add(asJson(lockedDependency, nodeFactory));
+  private JsonNode asJson(
+      Iterable<Dependency> lockedDependencies, JsonNodeFactory jsonNodeFactory) {
+    ArrayNode json = jsonNodeFactory.arrayNode();
+    for (Dependency lockedDependency : lockedDependencies) {
+      json.add(asJson(lockedDependency, jsonNodeFactory));
     }
     return json;
   }
-
-private JsonNode asJson(
-        Iterable<Dependency> lockedDependencies, JsonNodeFactory jsonNodeFactory) {
-    ArrayNode json = jsonNodeFactory.arrayNode();
-    for (Dependency lockedDependency : lockedDependencies) {
-        json.add(asJson(lockedDependency, jsonNodeFactory));
-    }
-    return json;
-}
 
   private JsonNode asJson(Dependency lockedDependency, JsonNodeFactory jsonNodeFactory) {
     ObjectNode json = jsonNodeFactory.objectNode();
@@ -87,7 +76,7 @@ private JsonNode asJson(
     json.put("artifactId", artifactIdentifier.artifactId);
     json.set(
         "version",
-        lockedDependency.getVersion().accept(versionConstraintJsonSerializer, nodeFactory));
+        lockedDependency.getVersion().accept(versionConstraintJsonSerializer, jsonNodeFactory));
     json.put("scope", lockedDependency.scope);
     json.put("type", artifactIdentifier.type);
     json.put("optional", lockedDependency.optional);
