+++ src/main/java/se/vandmo/dependencylock/maven/json/LockfileJson.java
@@ -38,21 +38,17 @@ import se.vandmo.dependencylock.maven.Parent;
 import se.vandmo.dependencylock.maven.Parents;
 import se.vandmo.dependencylock.maven.Plugin;
 import se.vandmo.dependencylock.maven.Plugins;
-import se.vandmo.dependencylock.maven.versions.VersionConstraint;
-import se.vandmo.dependencylock.maven.versions.VersionConstraints;
 import se.vandmo.dependencylock.maven.ProfileEntry;
 import se.vandmo.dependencylock.maven.ProfiledDependencies;
 
-public final class LockfileJson implements Lockfile {
+public final class LockfileJson extends WithJsonHelper implements Lockfile {
 
   private static final String V2 = "2";
   private static final String V3 = "3";
   private final LockFileAccessor dependenciesLockFile;
-  private final VersionConstraintJsonSerializer versionConstraintJsonSerializer;
 
   private LockfileJson(LockFileAccessor dependenciesLockFile) {
     this.dependenciesLockFile = dependenciesLockFile;
-    this.versionConstraintJsonSerializer = new VersionConstraintJsonSerializer();
   }
 
   public static LockfileJson from(LockFileAccessor dependenciesLockFile) {
@@ -241,22 +237,19 @@ public final class LockfileJson implements Lockfile {
   private JsonNode write(LockedProject contents, JsonNodeFactory jsonNodeFactory) {
     ObjectNode json = jsonNodeFactory.objectNode();
     json.put("version", V3);
-    json.set(
-        "artifacts", JsonUtils.buildArtifactsJson(collectArtifacts(contents), jsonNodeFactory));
+    json.set("artifacts", buildArtifactsJson(collectArtifacts(contents), jsonNodeFactory));
     contents.parents.ifPresent(parents -> json.set("parents", toJson(parents, jsonNodeFactory)));
     contents.plugins.ifPresent(plugins -> json.set("plugins", toJson(plugins, jsonNodeFactory)));
     contents.extensions.ifPresent(
         extensions -> json.set("extensions", toJson(extensions, jsonNodeFactory)));
     json.set(
         "dependencies",
-        JsonUtils.buildDependenciesJson(
-            contents.dependencies.getSharedDependencies(), jsonNodeFactory));
+        buildDependenciesJson(contents.dependencies.getSharedDependencies(), jsonNodeFactory));
     json.set(
-        "profiles",
-        JsonUtils.buildProfilesJson(contents.dependencies.profileEntries(), jsonNodeFactory));
+        "profiles", buildProfilesJson(contents.dependencies.profileEntries(), jsonNodeFactory));
     return json;
   }
-  
+
   private Stream<Artifact> collectArtifacts(LockedProject contents) {
     Stream<Artifact> result = contents.dependencies.artifacts();
     result = Stream.concat(result, contents.plugins.map(Plugins::artifacts).orElse(Stream.empty()));
